refactor: extract SSRF worker-URL validation into app/worker_proxy.py (CashPilot-sux)
Move the self-contained SSRF/worker-URL guard (policy constants, allowlist
parsing, IP classification, _validate_worker_url, _pin_url_to_ip) out of the
main.py god module into a focused, independently unit-testable app/worker_proxy.py.
main.py's _get_verified_worker_url composes them; it stays in main because it
also needs FLEET_API_KEY + database. worker_proxy imports nothing from main, so
no new import cycle is introduced. Behavior-preserving: 1172 tests green.

Test seams that referenced the moved symbols via app.main (the socket-getaddrinfo
patches and the strict-mode/allowlist tests mutating the policy constants) now
point at app.worker_proxy; the _validate_worker_url/_pin_url_to_ip patches keep
working through main's import binding.

This is the architect-endorsed least-coupled extraction from bead CashPilot-sux.
The benign, documented main->routers->main import cycle is intentionally left
as-is (resolving it risks the ~40 app.main.* test seams for no correctness gain).
